pyproject.toml
[project]
name = "glue-dispenser-twin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0.3",
    "rerun-sdk>=0.22.1",
]
